completed install script and various utils

+++ install.sh
@@ -1,15 +1,7 @@
 #!/bin/bash
 
+[[ -z $DEBUG ]] || set -x
 source `dirname $0`/scripts/common.sh
-
-download() {
-    url=$1
-    output=$2
-    [ -e $output ] && return 0
-    start_spinner "Downloading $url..."
-    Q wget `cat $url.url` -O $output
-    stop_spinner $?
-}
 
 usage() {
 cat << EOF
@@ -18,69 +10,86 @@ usage: $0 [<OPTIONS>]
 Installs and sets up buildroot for developing o11s with a pandaboard.
 
 OPTIONS:
-   -c   Make a clean build.
-   -v   Be verbose.
-   -h   Show this message.
+   -c         Make a clean build.
+   -C         Clean the target rootfs (THIS IS DANGEROUS)
+   -v         Be verbose.
+   -f         Force override of package installation and configuration files
+              on the host machine.
+   -d=<X,...> Disable certain install steps:
+               - nodownload: don't download anything
+               - noextract: don't extract anything
+               - nopackages: don't install support packages
+               - noconfig: don't generate configs
+               - nobuildroot: don't build buildroot
+               - noboot: don't boot the pandaboard
+               - noattach: don't attach an xterm to console of the pandaboard
+   -h         Show this message.
 EOF
 }
 
-Qorerr() {
-    [[ -n $VERBOSE ]] && ${*} || err=`${*} 2>&1`; ret=$?
-    [[ $ret -ne 0 ]] && echo $err
-    return $ret
-}
-
 # parse the incoming parameters
-while getopts "chv" options; do
+while getopts "ckChvfd:" options; do
     case $options in
 	c ) CLEAN=1;;
+	C ) CLEAN_TARGET=1;;
+	#k ) KILLSERIAL=1;;
     v ) VERBOSE=1;;
-    h ) usage exit 0;;
+    f ) FORCE=1;;
+    h ) usage; exit 0;;
+    d ) DISABLE=${OPTARG};;
     * ) echo unkown option: ${option}
         usage
         exit 1;;
     esac
 done
 
+h1 "Requesting root... "
+sudo echo "...got root."
+
 _pushd vendor
 
-download buildroot buildroot.tbz2
-download usbboot usbboot.deb
+if [[ ! $DISABLE =~ nodownload ]]; then
 
-start_spinner "Extracting buildroot..."
+    download BUILDROOT buildroot.tbz2
+    download USBBOOT usbboot.deb
 
-[[ -z $CLEAN ]] || Q rm -rvf $BUILDROOT || die "cleaning buildroot failed"
+fi
 
-Q mkdir -vp $BUILDROOT
+if [[ ! $DISABLE =~ noextract ]]; then
 
-[[ -d $BUILDROOT ]] || \
-    Q tar -xj -f buildroot.tbz2 \
-        --directory=$BUILDROOT \
-        --strip-components=1 || stop_spinner 1
+    start_spinner "Extracting buildroot..."
 
-stop_spinner 0
+    [[ -z $CLEAN ]] || Q rm -rvf $BUILDROOT || stop_spinner "cleaning buildroot failed"
 
-start_spinner "Extracting usbboot..."
+    Q mkdir -vp $BUILDROOT
 
-T=$OUT/usbboot
-[[ -z $CLEAN ]] || Q rm -rvf $T || die "cleaning usbboot failed"
+    [[ -d $BUILDROOT ]] || \
+        Q tar -xj -f buildroot.tbz2 \
+            --directory=$BUILDROOT \
+            --strip-components=1 || stop_spinner 1
 
-Q mkdir -vp $T
+    stop_spinner 0
 
-[[ -d $T ]] || \
-    Q dpkg -x usbboot.deb $T || stop_spinner 1
+    start_spinner "Extracting usbboot..."
 
-stop_spinner 0
+    T=$OUT/usbboot
+    [[ -z $CLEAN ]] || Q rm -rvf $T || die "cleaning usbboot failed"
 
-_popd
+    Q mkdir -vp $T
+
+    [[ -d $T ]] || \
+        Q dpkg -x usbboot.deb $T || stop_spinner 1
+
+    stop_spinner 0
+
+    _popd
+
+fi
 
 _pushd $BUILDROOT
 
-start_spinner "Patching buildroot..."
-
-Q $ROOT/scripts/patch_busybox
-
-stop_spinner 0
+h1 "Patching buildroot..."
+Q $ROOT/scripts/patch_buildroot || die "Failed to patch buildroot"
 
 if [[ -n $CLEAN ]]; then
 
@@ -93,8 +102,104 @@ if [[ -n $CLEAN ]]; then
     stop_spinner 0
 fi
 
-start_spinner "Downloading buildroot packages..."
-Qorerr make source
-stop_spinner $?
+if [[ ! $DISABLE =~ nodownload ]]; then
+
+    start_spinner "Downloading buildroot packages (takes a long time)..."
+    Qorerr make source
+    stop_spinner $?
+
+fi
 
 _popd
+
+if [[ ! $DISABLE =~ nopackages ]]; then
+
+    h1 "Installing required host packages..."
+
+    force_msg="(Use -f force override of existing packages)"
+
+    install_pkg udhcpd "Failed to install DHCP server: udhcpd! $force_msg" || exit 1
+    install_pkg nfs-kernel-server "Failed to install NFS server! $force_msg" || exit 1
+    install_pkg atftpd "Failed to install TFTP server: atftpd! $force_msg" || exit 1
+    install_pkg xinetd "Failed to install inet server: xinetd! $force_msg" || exit 1
+    install_pkg linaro-boot-utils "Failed to install linaro-boot-utils (for usbboot)" || exit 1
+
+fi
+
+if [[ ! $DISABLE =~ noconfig ]]; then
+
+    h1 "Setting up TFTP/NFS directory..."
+
+    sudo mkdir -p $NFS_ROOT
+    sudo mkdir -p $DEV_TARGET
+    sudo mkdir -p $TFTP_ROOT
+
+    h1 "Installing host config files..."
+    USB_ETH=$(find_asix_eth | item_at 1) 
+    [[ -n $USB_ETH ]] || exit 1
+
+    nfs_exports=$(sub_cfg_vars $HOST_OVERLAY/nfs_exports)
+    udhcpd_conf=$(sub_cfg_vars $HOST_OVERLAY/udhcpd.conf)
+    inetd_conf=$(sub_cfg_vars $HOST_OVERLAY/inetd.conf)
+
+    add_config_lines $nfs_exports /etc/exports
+    add_config_lines $udhcpd_conf /etc/udhcpd.conf
+    add_config_lines $inetd_conf /etc/inetd.conf
+
+    sudo mkdir -p $PXE_ROOT
+
+    usb_serial=$(find_usb_serial)
+    [[ -n $usb_serial ]] || exit 1
+
+    $SCRIPTS/boot_pandaboard || exit 1
+
+    h1 "Configuring usb interface for DHCP..."
+    Qorerr sudo ifconfig $USB_ETH $SERVER_IP
+    start_service udhcpd
+
+    h1 "Trying to figure out MAC addr for PXE (reading from $usb_serial)"
+    pxelin=$(find_pxe_mac_addr $usb_serial)
+    [[ -n $pxelin ]] || exit 1
+
+    h2 "Discovered PXE mac addr: $pxelin"
+    generate_pxe_config $pxelin
+
+fi
+
+if [[ ! $DISABLE =~ nobuildroot ]]; then
+
+    _pushd $BUILDROOT
+
+    start_spinner "Building buildroot..."
+    Qorerr make || exit 1
+    stop_spinner $?
+
+    _popd
+
+fi
+
+if [[ ! $DISABLE =~ noboot ]]; then
+
+    h1 "Booting the pandaboard..."
+
+    h2 "Starting required services..."
+
+    start_service nfs-kernel-server
+    start_service xinetd
+
+    $SCRIPTS/boot_pandaboard || exit 1
+fi
+
+if [[ ! $DISABLE =~ noattach ]]; then
+    if [[ -n $DISPLAY ]]; then
+        xterm -e "$SCRIPTS/attach_console" &
+    fi
+fi
+
+echo
+echo "PandaBoard dev environment should now be availabe..."
+echo
+echo "- SSH     : ssh test@192.168.99.2"
+echo "- Console : screen -rd pandaconsole"
+echo
+echo "Also see: scripts/attach_console"
